ml-model/requirements.txt
tensorflow==2.13.0
Flask==2.3.3
gunicorn==21.2.0
Pillow==10.0.0
numpy==1.24.3